use md5 as ketama hash and fnv1a32 to shard key
